object_detection/requirements.txt
# Object Detection Evaluation Requirements
# YOLO 物件偵測評測環境依賴

# Ultralytics YOLO - 主要物件偵測框架
ultralytics>=8.0.0

# PyTorch - YOLO 依賴
torch>=2.0.0
torchvision>=0.15.0

# 資料處理與分析
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0

# 圖像處理
Pillow>=10.0.0
scikit-image>=0.21.0

# YAML 配置檔處理
PyYAML>=6.0

# 進度條顯示
tqdm>=4.65.0

# 視覺化（選用）
matplotlib>=3.7.0
seaborn>=0.12.0

# RetinaFace for Dark Face Detection (Optional)
# git+https://github.com/biubug6/Pytorch_Retinaface.git
